refactor(#2957): extract maybeActivateAsync shared activation entry point (phase 1)
Byte-inert extraction (#2957 slice 1). The two async state-machine
activation blocks in compileFunctionBody (drive lane + CPS/host-drive
lane) plus the local rewriteFuncResultType helper move verbatim into a new
shared module src/codegen/async-activation.ts, exporting
maybeActivateAsync(ctx, fctx, decl, func): boolean. compileFunctionBody now
calls it. The internal ts.isFunctionDeclaration guards are preserved, so
declaration activation is unchanged; the decl param is widened to
ts.FunctionLikeDeclaration so phases 2-3 can wire the same entry point into
the arrow (closures.ts) and method (class-bodies.ts / literals.ts) paths.

Proof: compiling the playground+examples corpus (26 sources) x {gc, wasi,
standalone} plus all four async shapes yields the identical
CORPUS_SHA256=161cd89fb5a298fb86c76af6fcdcd787f42f340ba6ce988fecaf08cc78a18d4b
before and after (62 compiles, byte-for-byte matching). tsc --noEmit clean.
No behaviour change; phases 2-3 (the real arrow/method activation) remain open.

✓
